Create WAV of a Sine wave using Hound
Hound is a library that abstracts creating and reading WAV files with
Rust. This commit adds Hound to the project and introduces a function
`write_sine_wav` which uses the mathematical formula A.sin(2π𝒇𝒕) to
produce a sine wave.

Because we're using a computer and therefore must be using discrete
sampling we choose a sampling frequency (44.1kHz) and an appropriate
number of bits per sample (16), and we use the formula above to generate
88200 samples and write them to the wav file.

88200 samples at 44.1kHz (44100 samples per second) produces enough
points to build a 2 second long audio clip.

I picked 44.1kHz and 16 bits purely because they are the values used by
the CD-DA format used on audio CD's and defined by Sony and Phillips in
The Red Book in 1982[1]

[1]: https://en.wikipedia.org/wiki/Compact_Disc_Digital_Audio